requirements.txt
TikTokApi>=0.11.6
pytesseract>=0.3.10
Pillow>=9.5.0
aiohttp>=3.9.0
